requirements.txt
# Web Framework
flask>=2.3.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Template and Web Utils
jinja2>=3.1.0
markupsafe>=2.1.0

# Environment Variables
python-dotenv>=0.19.2

# Machine Learning, Embeddings, RAG
langchain>=0.1.0
faiss-cpu>=1.7.4
langchain-mistralai>=0.1.0
langchain-google-genai>=0.0.4

# Markdown
markdown>=3.4.0